==> 3_baseline_predictor.ipynb <==
# Baseline predictor
This baseline calculates the average remaining time for per index in the trace. For instance:
- 0 (i.e. first event) -> remaining time 20 days
- 1 (i.e. second event) -> remaining time 19 days
- etc.

# import basic libraries
import pandas as pd

# import machine learning libraries
from sklearn.metrics import mean_squared_error
from sklearn.metrics import r2_score

X_train = pd.read_csv("data/generated/onehot/X_train.csv")
y_train = pd.read_csv("data/generated/onehot/y_train.csv")
X_test = pd.read_csv("data/generated/onehot/X_test.csv")
y_test = pd.read_csv("data/generated/onehot/y_test.csv")

# merge X train and y train
train_df = result = pd.concat([X_train, y_train], axis=1)

# make a dictionary that maps an index to the average remaining time
avg_remaining_time_prediction = train_df.groupby('event_index_in_trace')['remaining_time'].mean().to_dict()
avg_remaining_time_prediction

## Make predictions

X_test['y_pred'] = X_test['event_index_in_trace'].map(avg_remaining_time_prediction)
X_test.head()

## Calculate metrics
The MSE on the baseline is around 100. This means that the predictions are off by about 10 days.

The R2 score of approximately 0.3 suggests that approximately 30% of the variance in the target variable is explained by the model. Ideally, we want our predictor to have a higher R2 score closer to 1.

# MSE
mean_squared_error(y_test["remaining_time"], X_test["y_pred"])

# R2
r2_score(y_test["remaining_time"], X_test["y_pred"])
