==> 4_catboost_stacked.ipynb <==
# Catboost predictor

# import basic libraries
import pandas as pd

# import machine learning libraries
from sklearn.metrics import mean_squared_error
from sklearn.metrics import r2_score
from sklearn.metrics import accuracy_score, precision_score, f1_score

#import catboost
from catboost import CatBoostRegressor

#import the data
X_train = pd.read_csv("data/generated/new_documents/X_train.csv")
y_train = pd.read_csv("data/generated/new_documents/y_train.csv")
X_test = pd.read_csv("data/generated/new_documents/X_test.csv")
y_test = pd.read_csv("data/generated/new_documents/y_test.csv")

X_test.head(50)

y_test = y_test['remaining_time']
y_train = y_train['remaining_time']

### Pre-process the data

# Convert timestamp to a pandas datetime object
X_train['timestamp'] = pd.to_datetime(X_train['time:timestamp'],format='ISO8601')

# Extract relevant features
X_train['year'] = X_train['timestamp'].dt.year
X_train['month'] = X_train['timestamp'].dt.month
X_train['day'] = X_train['timestamp'].dt.day
X_train['hour'] = X_train['timestamp'].dt.hour
X_train['minute'] = X_train['timestamp'].dt.minute
X_train['second'] = X_train['timestamp'].dt.second
X_train['microsecond'] = X_train['timestamp'].dt.microsecond 

# Drop the original timestamp column
X_train = X_train.drop(['time:timestamp', 'timestamp'], axis=1)

# Convert timestamp to a pandas datetime object
X_test['timestamp'] = pd.to_datetime(X_test['time:timestamp'],format='ISO8601')

# Extract relevant features
X_test['year'] = X_test['timestamp'].dt.year
X_test['month'] = X_test['timestamp'].dt.month
X_test['day'] = X_test['timestamp'].dt.day
X_test['hour'] = X_test['timestamp'].dt.hour
X_test['minute'] = X_test['timestamp'].dt.minute
X_test['second'] = X_test['timestamp'].dt.second
X_test['microsecond'] = X_test['timestamp'].dt.microsecond 

# Drop the original timestamp column
X_test = X_test.drop(['time:timestamp', 'timestamp'], axis=1)

# Drop the offer columns
X_train = X_train.drop(['FirstWithdrawalAmount', 'NumberOfTerms', 'Accepted','MonthlyCost','Selected','CreditScore','OfferedAmount','OfferID','Unnamed: 0'], axis=1)
# Drop the offer columns
X_test = X_test.drop(['FirstWithdrawalAmount', 'NumberOfTerms', 'Accepted','MonthlyCost','Selected','CreditScore','OfferedAmount','OfferID','Unnamed: 0'], axis=1)

#Drop some other unnecessary columns
#X_train = X_train.drop(['Action', 'EventID','case:ApplicationType','EventOrigin','lifecycle:transition'],axis=1)

#Drop some other unnecessary columns
#X_test = X_test.drop(['Action', 'EventID','case:ApplicationType','EventOrigin','lifecycle:transition'],axis=1)

#dropping the ID column
#X_train = X_train.drop(['case:concept:name'],axis=1)
#X_test = X_test.drop(['case:concept:name'],axis=1)

# Separate categorical and numeric features

categorical_features = ['org:resource','concept:name','case:LoanGoal','case:concept:name','Action', 'EventID','case:ApplicationType','EventOrigin','lifecycle:transition']
numeric_features = X_train.columns.difference(categorical_features)

# Convert categorical features to string
X_train[categorical_features] = X_train[categorical_features].astype(str)
X_test[categorical_features] = X_test[categorical_features].astype(str)

X_test.dtypes

### Building the model

# Create a CatBoost regressor
model = CatBoostRegressor(iterations=500, depth=6, learning_rate=0.01, loss_function='RMSE')

# Train the model
model.fit(X_train, y_train, cat_features=categorical_features, eval_set=(X_test, y_test), early_stopping_rounds=10, verbose=10)

# Make predictions on the test set
y_pred = model.predict(X_test)

### Evaluate the model

# MSE
mean_squared_error(y_test, y_pred)

# R2
r2_score(y_test, y_pred)

from sklearn.metrics import mean_squared_error, mean_absolute_error, r2_score, explained_variance_score

# Assuming y_test and y_pred are your actual and predicted values, respectively
mse = mean_squared_error(y_test, y_pred)
mae = mean_absolute_error(y_test, y_pred)
rmse = mean_squared_error(y_test, y_pred, squared=False)
r2 = r2_score(y_test, y_pred)
explained_variance = explained_variance_score(y_test, y_pred)

print(f"Mean Squared Error (MSE): {mse}")
print(f"Mean Absolute Error (MAE): {mae}")
print(f"Root Mean Squared Error (RMSE): {rmse}")
print(f"R-squared (R²): {r2}")
print(f"Explained Variance Score: {explained_variance}")

#Precision
precision = precision_score(y_test, y_pred, average='weighted')

#F1
f1 = f1_score(y_test, y_pred, average='weighted')

### Feature importance

import matplotlib.pyplot as plt

# Assuming 'model' is your trained CatBoostRegressor
feature_importance = model.get_feature_importance()

# Get feature names from the original DataFrame (X_train)
feature_names = X_train.columns

# Create a DataFrame to store feature importance values with their corresponding names
feature_importance_df = pd.DataFrame({'Feature': feature_names, 'Importance': feature_importance})

# Sort the DataFrame by importance in descending order
feature_importance_df = feature_importance_df.sort_values(by='Importance', ascending=False)

# Plot feature importance
plt.figure(figsize=(10, 6))
plt.barh(feature_importance_df['Feature'], feature_importance_df['Importance'])
plt.xlabel('Feature Importance')
plt.ylabel('Feature')
plt.title('CatBoost Feature Importance')
plt.show()

X_train.head()
